chore: move types to types.ts
